thermal: Add msm tsens thermal sensor driver
TSENS supports reading from multiple sensors and notifying
clients when the lower/upper thresholds are reached. The
lower/upper threshold use the last sensors threshold that are set
through the thermal sysfs.  TSENS HW is enabled only when the
main sensor is requested.  The TSENS block is disabled if the
main senors is disabled irrespective of any other sensors that
are being enabled.  When TSENS receives notification on
temperature threshold being reached, it checks and sends events
to the sensors that are enabled and whose thresholds are above
the upper or below the lower thresholds set in the tsens
threshold register.

Based on code by Siddartha Mohanadoss.

Cc: Siddartha Mohanadoss <smohanad@codeaurora.org>
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Narendran Rajan <nrajan@codeaurora.org>
Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>