clk: qcom: gdsc: handle dependencies with clocks
Some gdsc instances have dependencies with a clock which need to
be turned on before the gdsc itself can be turned on. one example
of such a dependency exist with some RPM clocks which need to be
voted, which inturn turn a RPM controlled gdsc on, before the APSS
controlled gdsc is turned on.

Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>