drm/msm/hdmi: Manage hdmi PLL through phy driver
Add a helper to initialize PLL in the phy driver.

HDMI PLLs are going to have their own mmio base different from that of
PHY.

For the clock code in hdmi_phy_8960.c, some changes were needed for it to
work with the updated register offsets. Create a copy of the updated clock
code in hdmi_pll_8960.c, instead of rewriting it in hdmi_phy_8960.c
itself. This removes the need to place CONFIG_COMMON_CLOCK checks all
around, makes the code more legible, and also removes some old checkpatch
warnings with the original code.

The older hdmi pll clock ops in hdmi_phy_8960.c will be removed later.

Signed-off-by: Archit Taneja <architt@codeaurora.org>