soc: qcom: rpmpd: Add a powerdomain driver to model cx/mx powerdomains
The cx/mx powerdomains just pass the performance state set by the
consumers to the RPM (Remote Power manager) which then takes care
of setting the appropriate voltage on the corresponding rails to
meet the performance needs.

Add data for all powerdomains on msm8996.

This is still a test driver and is not for merge.

Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>