mmc: sdhci-msm: Modify the driver to use OPP apis
Modify the driver to use OPP apis to set clock freq requirements instead
of using clk apis. To do this sepcify a OPP table in DT, specifiying the
performance states that needs to be set on the RPM powerdomain along with
setting the clock rates.

Revert back the custom floor ops that we done for sdhci.

Signed-off-by: Rajendra Nayak <rnayak@codeuarora.org>