arm64: dts: HACK: shuffle the clocks around for sdhci
OPP layer defaults to clocks with index 0 for device with OPP tables
to figure out what clock to scale. This might work for CPU devices,
but will not work for IO devices with multiple clocks.
Shuffle the iface and core clocks for now so we can get it working.
This needs to be fixed in the OPP layer.

Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>