clk: qcom: gdsc: Add support for registering gdscs via DT
To be able to describe a gdsc and rpm powerdomain hirerachy, we need
to have a way of registering gdscs which can be specified as child
nodes of the clock controller they belong to.
This allows for the gdscs that can then be hooked up to the parent
rpm powerdomains in DT. Add this support in the gdsc driver so it
registers all the DT provided gdscs and then falls back on the existing
way of registering gdscs based on driver data.

Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>