PM / Domain: Add struct device to genpd
The power-domain core would be using the OPP core going forward and the
OPP core has a basic requirement of a device structure for its working.

Add a struct device to the genpd structure and also add a genpd bus type
for the devices.

Note that the of_node field of the device is only set when separate DT
node is present for the power-domain, otherwise the of node is common
across multiple genpd devices and filling the of_node field with it
doesn't sound right.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>