mmc: sdhci-msm: Adapt the driver to use OPPs to set clocks/performance state
SDHCI driver needs to set a performance state along with scaling its
clocks. Modify the driver to use the newly introducded powerdomain
performance state based OPPs to scale clocks as well as set an
appropriate powerdomain performance state.

The patch also adds OPPs for sdhci device on msm8996.

The changes have to be validated by populating similar OPP tables
on all other devices which use the sdhci driver. This is for now
validated only on msm8996 and with missing OPP tables for other
devices is known to break those platforms.
Hence this patch for now is only for review and not for merge.

Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>