usb: phy: phy-msm-usb: Fix usage of devm_regulator_bulk_get()
The regulator_bulk_data pointer passed to devm_regulator_bulk_get()
is used to store the client handles for the regulators, which
then is used by devm_regulator_bulk_release() to free the
regulator.
Passing a local array as is done here means the memory used to
store the handles is freed causing the handles to be corrupted.

Fix this my moving the array inside of the msm_otg structure,
memory for which is dynamically allocated.

Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>