drivers: boot_constraint: Add support for supply constraints
This patch adds the first constraint type: power-supply.

The constraint is set by enabling the regulator and setting a voltage
range (if required) for the respective regulator device, which will be
honored by the regulator core even if more users turn up. Once the
device is probed, the regulator is released and the constraint is
removed.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>