drivers: boot_constraint: Add support for clk constraints
This patch adds the clk constraint type.

The constraint is set by enabling the clk for the device. Once the
device is probed, the clk is disabled and the constraint is removed.

We may want to do clk_set_rate() from here, but lets wait for some real
users that really want it.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>