drivers: boot_constraint: Add support for PM constraints
This patch adds the PM constraint type.

The constraint is set by attaching the power domain for the device,
which will also enable the power domain. This guarantees that the power
domain doesn't get shut down while being used.

We don't need to detach the power domain to remove the constraint as the
domain is attached only once, from here or before driver probe.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>