platform: msm: qcom_geni_se: don't return uninitalized value for probe
Variable 'ret' is unused in geni_se_probe() but it was being returned
at the bottom of the function anyway.   If it happened to be non-zero
then the system assumed that the subsystem didn't come up and other
stuff didn't work (like serial).  Return 0 for success and all is well.

Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>