PM / Domain: Implement of_dev_pm_genpd_get_performance_state()
This implements of_dev_pm_genpd_get_performance_state() which can be
used from the device drivers or the OPP core to find the performance
state encoded in the "required-opp" property of a node. Different
platforms may encode the performance state differently using the OPP
table (they may simply return value of opp-hz or opp-microvolt, or apply
some algorithm on top of those values) and so a new callback is
implemented to allow platform specific drivers to convert the power
domain OPP to a performance state.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>