clk: qcom: Add prepare/unprepare clock ops for PLL/RCG
To put across power domain votes associated with a PLL or RCG, add the
prepare/unprepare clock ops which would map the corresponding performance
state corners for a clock frequency when the clk_prepare/clk_unprepare is
being invoked.
Also update the set_rate clock ops to send across the performance state to
genpd framework when a new frequency is being requested.

Change-Id: Ic4999ddcd894755b972b8c1bd22b890803df680c
Signed-off-by: Taniya Das <tdas@codeaurora.org>